Add Neo4J for queries
Stardog needs a license and bump that. This proves that we need
to clean our triples so that we know that when it says he loves,
it means God or Jesus or whoever it means. But for now, this proves
that we can take some text, parse it for triples, process those
triples in some way using dataframes, and save this information inside
of a graph database for more research